Refactoracao tipos dos atributos ID's e Enums

+++ src/main/java/edu/uepb/web/biblioteca/model/Orientador.java
@@ -1,27 +1,29 @@
 package edu.uepb.web.biblioteca.model;
 
-
 /**
  * @autor geovanniovinhas <vinhasgeovannio@gmail.com
  *
  *
  */
 public class Orientador {
-	
-	private Long id;
+
+	private int id;
 	private String nome;
-	public Long getId() {
+
+	public int getId() {
 		return id;
 	}
-	public void setId(Long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getNome() {
 		return nome;
 	}
+
 	public void setNome(String nome) {
 		this.nome = nome;
 	}
 
 }
-
